+++ NEWS.md
@@ -1,6 +1,9 @@
 # targets 0.11.0.9000
 
+## Bug fixes
 
+* Fix `tar_mermaid()` graph ordering.
+* Hash the node names and quote the label names of `tar_mermaid()` graphs to avoid JavaScript keywords.
 
 # targets 0.11.0