+++ NEWS.md
@@ -1,4 +1,4 @@
-# targets 0.11.0.9000
+# targets 0.12.0
 
 ## Bug fixes